Refactor comparisons.
